Added rating of restaurants

+++ EasyBoni/Models/Restaurant.cs
@@ -23,6 +23,6 @@ namespace EasyBoni.Models
         [JsonProperty(PropertyName = "price")]
         public string Price { get; set; }
 
-        public float? Rating { get; set; }
+        public double? Rating { get; set; }
     }
 }
